+++ bm/routing.py
@@ -1,0 +1,13 @@
+
+from channels.auth import AuthMiddlewareStack
+from channels.routing import ProtocolTypeRouter, URLRouter
+
+from packages.routing import packages_ws_urlpatterns
+
+application = ProtocolTypeRouter({
+    'websocket': AuthMiddlewareStack(
+        URLRouter(
+            packages_ws_urlpatterns
+        )
+    )
+})