+++ packages/config.py
@@ -32,4 +32,14 @@ class PackageSettingsGeoloc:
 
     @staticmethod
     def interval_time():
-        return 60 #  equal to mins.
+        return 60 # equal to mins.
+
+    @staticmethod
+    def max_interval_time():
+        # mins
+        return 420
+
+    @staticmethod
+    def min_interval_time():
+        # mintus
+        return 20